Restructure benchmarks
Benchmarks named *ContinuousBenchmark should be executed regularly in a
benchmarking environment to track performance over time and to see if
there are regressions

To execute the continuous benchmarks, execute
./mvnw clean package -PskipTests=true
java -jar apm-agent-benchmarks/target/benchmarks.jar ".*ContinuousBenchmark" -prof gc
