+++ apm-agent-benchmarks/src/main/java/co/elastic/apm/benchmark/reporter/AbstractHttpJacksonReporterBenchmark.java
@@ -17,16 +17,21 @@
  * limitations under the License.
  * #L%
  */
-package co.elastic.apm.impl;
+package co.elastic.apm.benchmark.reporter;
 
-import co.elastic.apm.CpuProfiler;
+import co.elastic.apm.benchmark.profiler.CpuProfiler;
+import co.elastic.apm.report.serialize.JacksonPayloadSerializer;
+import co.elastic.apm.report.serialize.PayloadSerializer;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.module.afterburner.AfterburnerModule;
+import okio.Buffer;
+import org.openjdk.jmh.annotations.Setup;
 import org.openjdk.jmh.profile.GCProfiler;
 import org.openjdk.jmh.runner.Runner;
 import org.openjdk.jmh.runner.RunnerException;
 import org.openjdk.jmh.runner.options.OptionsBuilder;
 
-public class HttpVanillaJacksonReporterBenchmark extends AbstractHttpJacksonReporterBenchmark {
+public abstract class AbstractHttpJacksonReporterBenchmark extends AbstractHttpReporterBenchmark {
 
     /**
      * Convenience benchmark run method
@@ -36,14 +41,27 @@ public class HttpVanillaJacksonReporterBenchmark extends AbstractHttpJacksonRepo
      */
     public static void main(String[] args) throws RunnerException {
         new Runner(new OptionsBuilder()
-            .include(HttpVanillaJacksonReporterBenchmark.class.getSimpleName())
+            .include(AbstractHttpJacksonReporterBenchmark.class.getSimpleName())
             .addProfiler(GCProfiler.class)
             .addProfiler(CpuProfiler.class)
             .build())
             .run();
     }
 
-    protected ObjectMapper getObjectMapper() {
-        return new ObjectMapper();
+    @Setup
+    public void setUp() throws Exception {
+        super.setUp();
+        Buffer buffer = new Buffer();
+        getPayloadSerializer().serializePayload(buffer, payload);
+        System.out.println("Size of payload in bytes: " + buffer.readByteString().toByteArray().length);
     }
+
+    @Override
+    protected PayloadSerializer getPayloadSerializer() {
+        ObjectMapper objectMapper = getObjectMapper();
+        objectMapper.registerModule(new AfterburnerModule());
+        return new JacksonPayloadSerializer(objectMapper);
+    }
+
+    protected abstract ObjectMapper getObjectMapper();
 }
